subtraction demo
