+++ example/lib/globals.dart
@@ -1,6 +1,5 @@
 import 'dart:math';
 import 'package:fhel/seal.dart';
-import 'package:fhel_calculator/page/settings.dart';
 import 'package:flutter/material.dart';
 
 // ignore: constant_identifier_names
@@ -30,26 +29,6 @@ final Map noBatchingBV = {
   'ptMod': 4096,
   'secLevel': 128
 };
-
-/// Display a status banner and log [onFailure] message
-/// 
-/// If the operation is correct, display the actual result.
-/// Otherwise, display the failure message.
-ScaffoldFeatureController statusBanner(BuildContext context, SessionChanges session, bool isCorrect, String onSuccess, String onFailure) {
-  if (!isCorrect) {
-    session.log(onFailure);
-  }
-  return ScaffoldMessenger.of(context).showSnackBar(
-    SnackBar(
-      content: isCorrect
-        ? Text(onSuccess)
-        : Text(onFailure),
-      backgroundColor: isCorrect
-        ? successColor
-        : errorColor,
-    )
-  );
-}
 
 // Store the parameters that are used within the session
